load/requirements.txt
chardet
serene-metadata
python-libarchive
python-magic
backports.csv
